Re-work (potentially) how we create backups
Add a low-space check for /system during installation.
If low space is found, use /data to store the originals
and compress/tar them before moving them back at the end.

Update the uninstaller to support this backup method and
be able to restore from it.

Requires my updated busybox (for gzip, tar and awk support).